refactored function location in file structure
