+++ mobile/src/utils/navigationUtils.ts
